89 Fix CeleryPoller leaking exception across tasks
If CeleryPoller tried to get the result of a task that failed, the
exception would be raised, but the task would not be removed by
CeleryPoller. In the next task, the same exception would be raised
because CeleryPoller would try to get the result of the failed task
again, and the exception would be raised in the context of the second
request.

Fix this by making sure exceptions are raised in a saved stack context.